refine tests, extract common
